pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-telegram-bridge"
version = "0.1.0"
description = "Self-hosted Telegram bridge for Codex CLI with local web configuration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "pydantic>=2.8,<3",
  "python-telegram-bot>=22.0,<23",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9",
]

[project.scripts]
codex-telegram-bridge = "codex_telegram_bridge.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
